move functions from dbmodel
